SWARM-1014 - Quiet down boot logging
Motivation
----------

Too much useless output when booting.

Modifications
-------------

Silence the error-context message by making sure the management fraction
has the error context.

Silence some messages we see in our own testsuite due to @DefaultDeployment
scraping cdi.jar.  This is accopmlished by changing the package name of the
test itself.

Rev the fraction-plugin also, to remove detect classes from the fractions.

Result
------

Yet quieter logs.
