+++ fractions/wildfly/management/module.conf
@@ -1,2 +1,3 @@
 org.jboss.as.domain-management
 org.jboss.as.domain-http-interface
+org.jboss.as.domain-http-error-context
