+++ testsuite/testsuite-cdi/src/main/java/org/wildfly/swarm/cdi/test/Cheddar.java
@@ -1,4 +1,4 @@
-package org.wildfly.swarm.cdi;
+package org.wildfly.swarm.cdi.test;
 
 import javax.inject.Singleton;
